Added the post detail activity and layout. Changed the default click behavour ListPosts to show post details.
Changed the post handler so it can handle the different id's better, within a post resource, needs work still.
The post hander needs still is not able to deal with related posts well, the post at 0 is the post you want for the post detail anything after is a related post.